requirements.txt
django>=5.0.0,<5.1.0
djangorestframework>=3.14.0
pyyaml
django-cors-headers
